created new file spyre_test_utilities.py contains helper functions for creating module input generators
