+++ tests/spyre_test_base_common.py
@@ -51,6 +51,10 @@ from spyre_test_config_models import (
     SupportedModuleConfig,
     TestEntry,
 )
+from spyre_test_utilities import (
+    create_module_inputs_func_from_yaml,
+    create_module_inputs_func_from_config,
+)
 
 warnings.filterwarnings("ignore", category=pytest.PytestUnknownMarkWarning)
 
@@ -270,43 +274,6 @@ class TorchTestBase(PrivateUse1TestBase):  # type: ignore[name-defined]  # noqa:
                 )
                 continue
 
-            # Create a module input generator from the sample_inputs_func
-            def create_module_inputs_func(item):
-                def module_inputs_func(
-                    module_info, device, dtype, requires_grad, training, **kwargs
-                ):
-                    """Generated from YAML edits.modules.include"""
-                    try:
-                        from torch.testing._internal.common_modules import ModuleInput
-                        from torch.testing._internal.common_utils import FunctionInput
-                    except ImportError:
-                        return []
-
-                    # Use the build_module_input method if available
-                    if hasattr(item, "build_module_input"):
-                        test_device = (
-                            torch.device(device) if isinstance(device, str) else device
-                        )
-                        seed = kwargs.get("seed")
-                        return [
-                            item.build_module_input(
-                                seed=seed,
-                                test_device=test_device,
-                                FunctionInput=FunctionInput,
-                                ModuleInput=ModuleInput,
-                            )
-                        ]
-
-                    # Fallback: empty inputs
-                    return [
-                        ModuleInput(
-                            constructor_input=FunctionInput(),
-                            forward_input=FunctionInput(),
-                        )
-                    ]
-
-                return module_inputs_func
-
             # Create ModuleInfo and add to module_db
             try:
                 print(
@@ -314,7 +281,7 @@ class TorchTestBase(PrivateUse1TestBase):  # type: ignore[name-defined]  # noqa:
                 )
                 module_info = ModuleInfo(
                     module_cls,
-                    module_inputs_func=create_module_inputs_func(module_item),
+                    module_inputs_func=create_module_inputs_func_from_yaml(module_item),
                     skips=(),
                     decorators=None,
                     dtypes=(torch.float32, torch.float16),
@@ -365,71 +332,10 @@ class TorchTestBase(PrivateUse1TestBase):  # type: ignore[name-defined]  # noqa:
 
             module_info = matching_modules[0]
 
-            # Create a generator function that uses the inline input specs
-            def create_generator(config: SupportedModuleConfig):
-                def module_inputs_func(
-                    module_info, device, dtype, requires_grad, training, **kwargs
-                ):
-                    """Generated module input function from YAML config."""
-                    try:
-                        from torch.testing._internal.common_modules import ModuleInput
-                        from torch.testing._internal.common_utils import FunctionInput
-                    except ImportError:
-                        return []
-
-                    # Get seed from global config
-                    seed = kwargs.get("seed")
-                    test_device = (
-                        torch.device(device) if isinstance(device, str) else device
-                    )
-
-                    # Build constructor inputs
-                    constructor_spec = config.constructor_inputs
-                    if constructor_spec and constructor_spec.has_inputs():
-                        constructor_args = constructor_spec.build_cpu_args(
-                            seed=seed,
-                            op_name=module_info.name,
-                            test_device=test_device,
-                        )
-                        constructor_kwargs = constructor_spec.resolved_kwargs(
-                            test_device=test_device
-                        )
-                    else:
-                        constructor_args = []
-                        constructor_kwargs = {}
-
-                    constructor_input = FunctionInput(
-                        *constructor_args, **constructor_kwargs
-                    )
-
-                    # Build forward inputs
-                    forward_spec = config.forward_inputs
-                    if forward_spec and forward_spec.has_inputs():
-                        forward_args = forward_spec.build_cpu_args(
-                            seed=(None if seed is None else seed + 10000),
-                            op_name=module_info.name,
-                            test_device=test_device,
-                        )
-                        forward_kwargs = forward_spec.resolved_kwargs(
-                            test_device=test_device
-                        )
-                    else:
-                        forward_args = []
-                        forward_kwargs = {}
-
-                    forward_input = FunctionInput(*forward_args, **forward_kwargs)
-
-                    return [
-                        ModuleInput(
-                            constructor_input=constructor_input,
-                            forward_input=forward_input,
-                        )
-                    ]
-
-                return module_inputs_func
-
             # Replace the module's input generator
-            module_info.module_inputs_func = create_generator(module_config)
+            module_info.module_inputs_func = create_module_inputs_func_from_config(
+                module_config
+            )
 
     @classmethod
     def _should_run(
